[Fix] Comit Before Merge

+++ PushPop/Assets/3.Script/04_Seoyoung/ETC/DataManager2.cs
@@ -53,13 +53,6 @@ public class BadWord
     public string badword;
 }
 
-
-[System.Serializable]
-public class Vulgarism
-{
-    public int index;
-    public string vulgarism;
-}
 #endregion
 
 
@@ -72,36 +65,30 @@ public class DataManager2 : MonoBehaviour
     public List<CategoryDict> categoryDicts_List = new List<CategoryDict>();
     public List<Scripts> scipts_List = new List<Scripts>();
     public List<IconDict> iconDicts_List = new List<IconDict>();
-
-
-    //일단 임시로 쓰는 딕셔너리..나중에 Mold_Dictionary의 딕셔너리로 변경
+    //파싱한 정보 저장할 딕셔너리
     public Dictionary<int, string> categoryDict = new Dictionary<int, string>();
     public Dictionary<int, string> iconDict = new Dictionary<int, string>();
+
 
     //도움말 스크립트용
     public List<HelpScript> helpScripts_List = new List<HelpScript>();
 
 
-
     //욕설방지용
-    public BadWord[] badWord_Arr;
+    public BadWord[] badWord_Arr;      //vulgarsim.json 파일이 켜는순간 컴퓨터가 멈춰서 일단 따로
     public string[] vulgarism_Arr;
 
 
-
-    //파일 이름
+    //파일 이름 (이 친구들은 Assets/StreamingAssets 폴더에 미리 담겨져 있어야 합니다)
     public string categoryDict_fileName = "category.json";
     public string iconDict_fileName = "icon.json";
     public string helpScript_fileName = "help.json";
     public string badWord_fileName = "badword.json";
     public string vulgarism_fileName = "vulgarism.json";
 
-    public string Datapath = string.Empty;
-
-
-
-
-    private string path = Application.streamingAssetsPath;
+    //경로
+    public string Datapath = string.Empty;      //persistentDataPath
+    public string path = string.Empty;      //StreamingDataPath;
 
 
     #region Unity Callback
@@ -119,27 +106,23 @@ public class DataManager2 : MonoBehaviour
         }
 
 
-
         Datapath = Application.persistentDataPath + "/gameData";
-        if (!File.Exists(Datapath)) // 경로 탐색
+        path = Application.streamingAssetsPath;
+
+        //폴더 생성
+        if (!File.Exists(Datapath))
         {
             Directory.CreateDirectory(Datapath);
         }
 
-        Save_HelpScript();
-        Read_HelpScript();
-
-        Read_Category();
-        Read_Icon();
-        Read_BadWord();
-        Read_Vulgarism();
+        Read_HelpScript();  //도움말 읽기
+        Read_Category();    //카테고리 이름 정보 -> 딕셔너리(숫자,한글)
+        Read_Icon();    //몰드 아이콘 이름 정보 -> 딕셔너리(숫자, 한글)
+        Read_BadWord();     //비속어 1
+        Read_Vulgarism();   //비속어 2
 
     }
 
-    private void Start()
-    {
-
-    }
     #endregion
 
 
@@ -157,27 +140,30 @@ public class DataManager2 : MonoBehaviour
 
     //category.json 읽어와서 Dictionary로 변환하는 메소드
     public void Read_Category()
-    {
+    {//category.json파일 읽어와 persistentDataPath에 저장 밑 json데이터 Dictionary로 파싱하는 메소드   
         //1. JSON 데이터를 Category형 List로 받아옴
         //2. List를 Dictinory로 변환
         //categoryDict.Clear();
 
         string oriPath = Path.Combine(path, categoryDict_fileName);
+        string realPath = Datapath + "/" + categoryDict_fileName;
 
-        WWW reader = new WWW(oriPath);
-        while (!reader.isDone)
+        if(!File.Exists(realPath))
         {
 
+            WWW reader = new WWW(oriPath);
+            while (!reader.isDone)
+            {
+
+            }
+
+            byte[] data = reader.bytes;
+            string resultData = System.Text.Encoding.UTF8.GetString(data);
+
+            //File.WriteAllBytes(realPath, reader.bytes);
+            File.WriteAllText(realPath, resultData);
         }
 
-        string realPath = Datapath + "/category.json";
-
-
-        byte[] data = reader.bytes;
-        string resultData = System.Text.Encoding.UTF8.GetString(data);
-
-        //File.WriteAllBytes(realPath, reader.bytes);
-        File.WriteAllText(realPath, resultData);
 
         string JsonString = File.ReadAllText(realPath);
         JsonData jsonData = JsonMapper.ToObject(JsonString);
@@ -207,7 +193,6 @@ public class DataManager2 : MonoBehaviour
             ca.number = item.Key;
             ca.name = item.Value;
             categoryDicts_List.Add(ca);
-
         }
 
         #region 파일/폴더 체크
@@ -225,41 +210,30 @@ public class DataManager2 : MonoBehaviour
 
         #endregion
 
-
-
-
         JsonData jsonData = JsonMapper.ToJson(categoryDicts_List);
-
-        //byte[] b = System.Text.UTF8Encoding.UTF8.GetBytes(jsonData.ToString());
-
-        //File.WriteAllBytes(path + "/" + categoryDict_fileName, b);
         File.WriteAllText(path + "/" + categoryDict_fileName, Decode_EncodedNonASCIICharacters(jsonData.ToString()));
     }
 
     //icon.json 읽어와서 Dictionary로 변환하는 메소드
     public void Read_Icon()
-    {
-        //iconDict.Clear();
-
-        #region 이부분을 처음 깔때만 실행하면 참 좋을텐데
+    {//icon.json파일 읽어와 persistentDataPath에 저장 밑 json데이터 Dictionary로 파싱하는 메소드      
         string oriPath = Path.Combine(path, iconDict_fileName);
-
-        WWW reader = new WWW(oriPath);
-        while (!reader.isDone)
-        {
-
-        }
-
         string realPath = Datapath + "/" + iconDict_fileName;
 
-        byte[] data = reader.bytes;
-        string resultData = System.Text.Encoding.UTF8.GetString(data);
+        if(!File.Exists(realPath))
+        {
+            WWW reader = new WWW(oriPath);
+            while (!reader.isDone)
+            {
 
+            }
 
-        //File.WriteAllBytes(realPath, reader.bytes);
-        File.WriteAllText(realPath, resultData);
+            byte[] data = reader.bytes;
+            string resultData = System.Text.Encoding.UTF8.GetString(data);
 
-        #endregion
+            File.WriteAllText(realPath, resultData);
+
+        }
 
         string JsonString = File.ReadAllText(realPath);
 
@@ -268,7 +242,6 @@ public class DataManager2 : MonoBehaviour
         for (int i = 0; i < jsonData.Count; i++)
         {
             iconDict.Add(int.Parse(jsonData[i]["number"].ToString()), jsonData[i]["name"].ToString());
-
         }
 
     }
@@ -330,60 +303,77 @@ public class DataManager2 : MonoBehaviour
     }
 
     public void Read_HelpScript()
-    {
-        //도움말 스크립트 읽어오는 함수
+    {//helpscript.json파일 읽어와 persistentDataPath에 저장 밑 json데이터를 파싱하는 메소드
 
+        string oriPath = Path.Combine(Application.streamingAssetsPath, helpScript_fileName);
+        string realPath = Datapath + "/" + helpScript_fileName;
 
-        if (File.Exists(Datapath + "/help.json"))
+        if(!File.Exists(realPath))
         {
-            Debug.Log("Read_HelpScript 파일 존재");
-            helpScripts_List.Clear();
-            string JsonString = File.ReadAllText(Datapath + "/help.json");
-            var jsonData = JsonMapper.ToObject(JsonString);
-
-            for (int i = 0; i < jsonData.Count; i++)
+            WWW reader = new WWW(oriPath);
+            while (!reader.isDone)
             {
-                HelpScript helpScript = new HelpScript();
-
-                helpScript.type = (string)jsonData[i]["type"];
-
-                //Debug.Log(jsonData[i]["script"]); //Json Data Array로 나옴 -> 안풀림..
-
-                for (int j = 0; j < jsonData[i]["script"].Count; j++)
-                {
-                    JsonData item = JsonMapper.ToObject<JsonData>(jsonData[i]["script"].ToJson());
-                    Scripts script = new Scripts();
-                    script.pageNum = int.Parse(item[j]["page"].ToString());
-                    script.content = item[j]["content"].ToString();
-
-                    helpScript.script.Add(script);
-                    //Debug.Log(script.pageNum + "," + script.content);
-                }
-
-                helpScripts_List.Add(helpScript);
+            
             }
+
+            byte[] data = reader.bytes;
+            string resultData = System.Text.Encoding.UTF8.GetString(data);
+
+            //File.WriteAllBytes(realPath, reader.bytes);
+
+            File.WriteAllText(realPath, resultData);
         }
+
+
+        helpScripts_List.Clear();
+        string JsonString = File.ReadAllText(Datapath + "/help.json");
+        var jsonData = JsonMapper.ToObject(JsonString);
+
+        for (int i = 0; i < jsonData.Count; i++)
+        {
+            HelpScript helpScript = new HelpScript();
+
+            helpScript.type = (string)jsonData[i]["type"];
+
+            //Debug.Log(jsonData[i]["script"]); //Json Data Array로 나옴 -> 안풀림..
+
+            for (int j = 0; j < jsonData[i]["script"].Count; j++)
+            {
+                JsonData item = JsonMapper.ToObject<JsonData>(jsonData[i]["script"].ToJson());
+                Scripts script = new Scripts();
+                script.pageNum = int.Parse(item[j]["page"].ToString());
+                script.content = item[j]["content"].ToString();
+
+                helpScript.script.Add(script);     
+            }
+
+            helpScripts_List.Add(helpScript);
+        }
+
 
     }
 
     public void Read_BadWord()
-    {
+    {//badword.json파일 읽어와 persistentDataPath에 저장 및 json데이터 배열로 파싱하는 메소드
+
         string oriPath = Path.Combine(path, badWord_fileName);
-
-        WWW reader = new WWW(oriPath);
-        while (!reader.isDone)
-        {
-
-        }
-
         string realPath = Datapath + "/" + badWord_fileName;
 
-        byte[] data = reader.bytes;
-        string resultData = System.Text.Encoding.UTF8.GetString(data);
+        if (!File.Exists(realPath))
+        {
 
+            WWW reader = new WWW(oriPath);
+            while (!reader.isDone)
+            {
 
-        File.WriteAllText(realPath, resultData);
+            }
 
+            byte[] data = reader.bytes;
+            string resultData = System.Text.Encoding.UTF8.GetString(data);
+
+            File.WriteAllText(realPath, resultData);
+        }
+ 
 
         string JsonString = File.ReadAllText(realPath);
 
@@ -396,9 +386,7 @@ public class DataManager2 : MonoBehaviour
             BadWord badword = new BadWord();
 
             badword.index = i;
-            badword.badword = jsonData[i]["badword"].ToString();
-
-           
+            badword.badword = jsonData[i]["badword"].ToString();           
             badWord_Arr[i] = badword;
         }
 
@@ -406,32 +394,36 @@ public class DataManager2 : MonoBehaviour
 
 
     public void Read_Vulgarism()
-    {
-        string oriPath = Path.Combine(path, vulgarism_fileName);
+    {//vulgarism.json파일 읽어와 persistentDataPath에 저장 및 json데이터를 배열로 파싱하는 메소드
 
-        WWW reader = new WWW(oriPath);
-        while (!reader.isDone)
+        string oriPath = Path.Combine(path, vulgarism_fileName);
+        string realPath = Datapath + "/" + vulgarism_fileName;
+
+        if(!File.Exists(realPath))
         {
+
+            WWW reader = new WWW(oriPath);
+            while (!reader.isDone)
+            {
+
+            }
+            byte[] data = reader.bytes;
+            string resultData = System.Text.Encoding.UTF8.GetString(data);
+            File.WriteAllText(realPath, resultData);
 
         }
 
-        string realPath = Datapath + "/" + vulgarism_fileName;
-
-        byte[] data = reader.bytes;
-        string resultData = System.Text.Encoding.UTF8.GetString(data);
-
-        File.WriteAllText(realPath, resultData);
-
         string JsonString = File.ReadAllText(realPath);
-
         JsonData jsonData = JsonMapper.ToObject(JsonString);
 
-        //vulgarisms_List
-
-
         string oneData = jsonData[0]["vulgarism"].ToString();
-
         vulgarism_Arr = oneData.Split(",");
+
+        for (int i =0; i<vulgarism_Arr.Length; i++)
+        {
+            vulgarism_Arr[i].Replace(" ", ""); 
+        }
+
 
     }
